[SE2R-42] Mappers tested + authorization in controllers

+++ backend/src/main/java/com/recipes/backend/rest/RecipesController.java
@@ -1,12 +1,14 @@
 package com.recipes.backend.rest;
 
 import com.recipes.backend.bizz.recipe.RecipeService;
+import com.recipes.backend.bizz.security.SecurityService;
 import com.recipes.backend.exception.domain.RecipeEmptyException;
 import com.recipes.backend.mapper.RecipeMapper;
 import com.recipes.backend.rest.domain.RecipeAllRest;
 import com.recipes.backend.rest.domain.RecipeRest;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpHeaders;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -14,22 +16,35 @@ import java.util.Optional;
 import java.util.Set;
 import java.util.stream.Collectors;
 
+import static com.recipes.backend.utils.LogWriter.logHeaders;
+
 @RestController
 @RequestMapping(path = "/recipes")
-public class RecipesController {
+public class RecipesController
+{
 
     private final RecipeService recipeService;
+    private final SecurityService securityService;
 
     @Autowired
-    public RecipesController(final RecipeService recipeService) {
+    public RecipesController(final RecipeService recipeService,
+                             final SecurityService securityService)
+    {
         this.recipeService = recipeService;
+        this.securityService = securityService;
     }
 
     @GetMapping
     public ResponseEntity<RecipeAllRest> getAllRecipes(@RequestHeader HttpHeaders headers,
                                                        @RequestParam(name = "page") int page,
-                                                       @RequestParam(name = "limit") int limit) {
-        //TODO check headers
+                                                       @RequestParam(name = "limit") int limit)
+    {
+        logHeaders(headers);
+
+        if (!securityService.isAuthenticated(headers))
+        {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
 
         final Set<RecipeRest> retrievedRecipes =
                 recipeService.getAllRecipes(page, limit).stream()
@@ -44,8 +59,14 @@ public class RecipesController {
 
     @DeleteMapping("/{id}")
     public ResponseEntity<String> deleteRecipe(@RequestHeader HttpHeaders headers,
-                                               @PathVariable(name = "id") Long recipeId) {
-        //TODO add header validation
+                                               @PathVariable(name = "id") Long recipeId)
+    {
+        logHeaders(headers);
+
+        if (!securityService.isAuthenticated(headers))
+        {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
         return recipeService.deleteRecipe(recipeId) ? ResponseEntity.ok(recipeId.toString()) : ResponseEntity.badRequest().body("Bad request!");
     }
 
@@ -54,8 +75,13 @@ public class RecipesController {
                                                @PathVariable(name = "id") Long recipeId,
                                                @RequestBody RecipeRest recipeRest)
     {
-        recipeService.updateRecipe(RecipeMapper.mapToRecipe(recipeRest).orElseThrow(RecipeEmptyException::new));
+        logHeaders(headers);
 
+        if (!securityService.isAuthenticated(headers)) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
+
+        recipeService.updateRecipe(RecipeMapper.mapToRecipe(recipeRest).orElseThrow(RecipeEmptyException::new));
         return ResponseEntity.ok().build();
     }
 }
